Best practice fixes for using pip
You shouldn't run pip as root unless you need to (eg docker container
that doesn't have a user account). This commit recommends running pip as
a user with the `--user` flag.  Not that in the docker (root only)
situation, using the `--user` flag will install stuff, but doesn't
appear to add it to your path, so you will "have" cmake, but can't
easily use it.