Introduce new Long Term Support (LTS) policy
Also adjustments to the existing End of Life (EOL) policy.

The new policy is:

- Non-LTS feature releases will be supported with critical patch
  versions for *at least* **four (4) months** from the initial
  publication date of the ***next*** feature release or until the
  *next*-***next*** feature release is published.

- Non-LTS feature releases will be allowed access to download
  signatures until at least four (4) months after the *next*-
  ***next*** feature release is published.

The EOL when the next-next feature release is published was previously
implied, and is now explicit.

But this is a reduction from 6 months to 4 months to match the number
of months in the CVD download access policy so as to reduce confusion.

The ClamAV team intends to increase feature release cadence and
so the effective number of supported versions, given the EOL after a
next-next release, should be the same.