PCRE sigs: Escaping forward slash is optional
The documentation claims that escaping forward slashes in PCRE
subsignatures is required and implies this is because `/` is used as the
delimiter for splitting the PCRE subsignature. 
This isn't true. ClamAV find the start and end of the PCRE subsig
by searching from the beginning for the first `/` and from the end
for the last `/`. 
So, escaping the forward slash (e.g. "`\/'") is optional. 
In testing, I found that the PCRE library will ignore any escape on `/`.